6-8까지 풀이

+++ practice/ch06/daewon/daewon_5.cpp
@@ -1,5 +1,4 @@
 #include <iostream>
-#include <string>
 using namespace std;
 
 class ArrayUtility {
@@ -11,12 +10,12 @@ public:
 
 void ArrayUtility::intToDouble(int source[], double dest[], int size)
 {
-    for(int i = 0; i < size; i++) dest[i] = source[i];
+    for(int i = 0; i < size; i++) dest[i] = (double)source[i];
 }
 
 void ArrayUtility::doubleToInt(double source[], int dest[], int size)
 {
-    for(int i = 0; i < size; i++) dest[i] = source[i];
+    for(int i = 0; i < size; i++) dest[i] = (int)source[i];
 }
